django-users/requirements.txt
asgiref==3.3.4
autopep8==1.5.7
billiard==3.6.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
decorator==4.4.2
Django==3.2
django-admin-footer==1.0.1
django-admin-interface==0.14.2
django-admin-logs==1.0.2
django-appconf==1.0.4
django-apptemplates==1.5
django-cors-headers==3.7.0
django-admin-honeypot==1.1.0
django-defender==0.9.1
django-extensions==3.1.0
django-filter==2.4.0
django-postgres-metrics==0.10.1
django-redis==4.12.1
django-request==1.5.6
django-safedelete==0.5.6
django-simple-history==3.0.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
docutils==0.16
Faker==5.6.0
future==0.18.2
Markdown==3.3.3
MarkupSafe==2.0.1
Pillow==8.1.0
proglog==0.1.9
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
pycodestyle==2.7.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.5
redis==3.5.3
requests==2.25.1
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.3
vine==1.3.0
django-cachalot